Move helper methods to the ApigeeMockApiClientHelperTrait
