+++ modules/apigee_edge_actions/tests/src/Kernel/Plugin/RulesEvent/EdgeEntityEventTestBase.php
@@ -20,12 +20,6 @@
 
 namespace Drupal\Tests\apigee_edge_actions\Kernel\Plugin\RulesEvent;
 
-use Apigee\Edge\Api\Management\Entity\App;
-use Drupal\apigee_edge\Entity\DeveloperApp;
-use Drupal\apigee_edge\Entity\DeveloperAppInterface;
-use Drupal\apigee_edge\Entity\EdgeEntityInterface;
-use Drupal\apigee_edge_teams\Entity\Team;
-use Drupal\apigee_edge_teams\Entity\TeamInterface;
 use Drupal\Core\Database\Database;
 use Drupal\Tests\apigee_mock_api_client\Traits\ApigeeMockApiClientHelperTrait;
 use Drupal\Tests\rules\Kernel\RulesKernelTestBase;
@@ -93,64 +87,6 @@ class EdgeEntityEventTestBase extends RulesKernelTestBase {
     ]);
     $this->account->save();
     $this->queueDeveloperResponse($this->account, Response::HTTP_CREATED);
-  }
-
-  /**
-   * Helper to create a DeveloperApp entity.
-   *
-   * @return \Drupal\apigee_edge\Entity\DeveloperAppInterface
-   *   A DeveloperApp entity.
-   *
-   * @throws \Drupal\Core\Entity\EntityStorageException
-   */
-  protected function createDeveloperApp(): DeveloperAppInterface {
-    /** @var \Drupal\apigee_edge\Entity\DeveloperAppInterface $entity */
-    $entity = DeveloperApp::create([
-      'appId' => 1,
-      'name' => $this->randomMachineName(),
-      'status' => App::STATUS_APPROVED,
-      'displayName' => $this->randomMachineName(),
-    ]);
-    $entity->setOwner($this->account);
-    $this->queueDeveloperAppResponse($entity);
-    $entity->save();
-
-    return $entity;
-  }
-
-  /**
-   * Helper to create a Team entity.
-   *
-   * @return \Drupal\apigee_edge_teams\Entity\TeamInterface
-   *   A Team entity.
-   *
-   * @throws \Drupal\Core\Entity\EntityStorageException
-   */
-  protected function createTeam(): TeamInterface {
-    /** @var \Drupal\apigee_edge_teams\Entity\TeamInterface $team */
-    $team = Team::create([
-      'name' => $this->randomMachineName(),
-      'displayName' => $this->randomGenerator->name(),
-    ]);
-    $this->queueCompanyResponse($team->decorated());
-    $this->queueDeveloperResponse($this->account);
-    $team->save();
-
-    return $team;
-  }
-
-  /**
-   * Helper to add Edge entity response to stack.
-   *
-   * @param \Drupal\apigee_edge\Entity\EdgeEntityInterface $entity
-   *   The Edge entity.
-   */
-  protected function queueDeveloperAppResponse(EdgeEntityInterface $entity) {
-    $this->stack->queueMockResponse([
-      'get_developer_app' => [
-        'app' => $entity,
-      ],
-    ]);
   }
 
   /**
